requirements.txt
numpy
pandas
PyMySQL
SQLAlchemy==1.4.49
certifi
charset-normalizer
idna
requests
urllib3
python-bugzilla==3.2.0
atlassian-python-api==3.41.19
PyYAML
beautifulsoup4
jinja2
